Add upgrade-setup-root.service/target
This patch introduces upgrade-setup-root.service, which enters the
target system (via chroot or other means) and sets up the disks for the
upgrade.

If the upgrade flags are in place, upgrade-pre-pivot.sh will remove
/run/initramfs/switch-root.conf (which disables the switch-root) and
then isolate upgrade-setup-root.target.

After upgrade-setup-root.target is upgrade-setup-root.service, which
sets up the disks in the root device for us. Once it's finished, it
isolates upgrade.target, which should start the upgrade process as
before.