Make kernel modules available in old root
We might still need to load kernel modules in the old root - for
example, if we need to mount /boot/efi, we'll need vfat, which might not
already be loaded.

This patch bind-mounts /lib/modules/$(uname -r) into the target system
so we can find and install them as needed.