Don't delete symlink before upgrade
Originally we were going to use the symlink as a flag to run the
upgrade, so we had to delete it immediately to prevent reboot loops.

But since we need to add a boot entry anyway for the newer
kernel/initramfs, we can put something like 'upgrade' on the boot
commandline to decide whether to run in upgrade mode.

Therefore: we can leave this symlink in place so other stuff can use it
later.