Don't switch plymouth mode until the upgrade starts (#873144)
We were switching the plymouth mode to "updates" because it stopped the
progress meter. But plymouth doesn't like doing update-root-fs if mode
!= boot.

If you don't do update-root-fs, plymouth doesn't know where
to find /proc or /dev or its plugins, so there's no connection to the
console (hence no text output) and it can't find the graphical plugin
to start it back up.

So: use pause-progress to stop the progress meter, and then switch to
"updates" mode once the upgrade starts.