fix plymouthd crash with encrypted /home (#896023)
When we switch into the old system to set up disks, our (new) plymouthd
keeps running and switches along with us. If the user has an encrypted
/home (or anything else that isn't / or /usr), plymouthd will try to
load modules so it can bring up the passphrase dialog.

But since we're now running in the *old* system, the modules might have
a different ABI from the server. And if that happens, plymouthd crashes,
you never get the unlock dialog, and the upgrade just hangs. Boo.

So here is a slightly grody workaround: bind-mount plymouth's module dir
over the old system's module dir, the same way we do for kernel modules.
Then plymouthd should be able to load everything fine and the rest
should Just Work, as usual.