write upgrade state to /var/tmp/upgrade.state
Currently this writes one of three strings which indicate the last known
status of the 'upgrade' hook:

  "running":  the upgrade hook has started.
  "finished": the upgrade hook returned an exit code of zero.
  "failed":   the upgrade hook returned a non-zero exit code.

This file can be used by upgrade-post scripts to react to an upgrade
failure, and it can be used by the system to report success/failure
after a reboot. It should be removed promptly when the system restarts
to avoid keeping stale state around.