Add initrd-system-upgrade-generator
This adds a systemd generator for the second phase of a system upgrade;
if /sysroot/system-upgrade is present (and a symlink to a directory)
when systemd starts, then we're back in the initramfs for upgrade time.

In that case, switch the default target to 'upgrade.target' so we'll run
the upgrade.