Prepare older filesystems before upgrade
In the pre-pivot hook, ensure that /run exists in the filesystem root so
that systemd can bind the initramfs /run mount into it. Remount /sysroot
with the original options (ro) after new directories are created so that
the post-pivot fsck doesn't get upset.

In the upgrade-pre hook, run convertfs on unconverted sysroots.