Skip udev cleanup when upgrading from non-systemd
If the system being upgraded from does not use systemd, mask the
initrd-udevadm-cleanup-db service. Normally, the handling of udev data
would happen as follows:

  - systemd in the initramfs runs systemd-udevd.service, which writes
    data about devices to /run/udev/data
  - systemd in the initramfs runs initrd-udevadm-cleanup-db, which
    deletes most of that data
  - switch root
  - systemd in the real root runs systemd-udevd.service with the real
    root hooks and rules, which writes data in /run/udev/data again

When upgrading from RHEL6, there is not systemd in the real root, so
nothing repopulates /run/udev/data. This data is needed by dracut when
generating the initramfs for the new kernel, so for this case just leave
behind the data from the initramfs.