Always put shell on tty2
Sometimes people get kinda freaked out if the upgrade seems like it's
not doing anything (like, say, maybe it's taking 90 seconds to run a
package %post script or something). And if they can't switch to a shell
and confirm that the system is actually doing stuff, they assume it's
hung, and reboot, and totally destroy their system.

So: always put a shell on tty2. And make sure the shell gets restarted
when we switch roots, so it actually works like it's supposed to.

If you really don't want the shell there for whatever reason, you can
add 'rd.upgrade.noshell' to the boot arguments to suppress it.