Fix backward compatibility with fedup-0.7.x
Fedup 0.7.x still exists in the world, and it has some some special
needs in order for us to get the system to start the upgrade:

1) /run/initramfs/upgrade.conf needs to exist and contain the usual
values for UPGRADEROOT and UPGRADELINK.

2) /proc/cmdline will contain "systemd.unit=system-upgrade.target", so
we need to make sure that exists inside the initramfs.

So let's set that stuff up, so we don't stick people in infinite reboot
loops if they're using an old version of fedup with a newer
fedup-dracut.