Don't write the entire system journal to upgrade.log
Prior to fedup 0.9.x, the upgrade process kept its journal in-memory
only, with a different machine-id, so if we wanted a text version of the
upgrade log you could just dump the entire journal to a file.

With fedup 0.9.x the upgrade logs go to your system's on-disk journal,
which means dumping the entire journal dumps your entire system log for
as far back as that goes. Whoops.

This just adds '-b' to the journalctl call so we only get logs for the
current boot.

(Bonus cleanup: make certain that our shell does the right thing with
the backup logfile name by doing "${logfile}.old")

Resolves: rhbz#1161366